@@ -8,14 +8,14 @@
 package com.facebook.stetho.common.android;
 
 import android.app.Activity;
-import android.os.Build;
+
 import com.facebook.stetho.common.ReflectionUtil;
-import com.facebook.stetho.common.Util;
+
+import java.lang.reflect.Field;
+import java.util.List;
 
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.NotThreadSafe;
-import java.lang.reflect.Field;
-import java.util.List;
 
 /**
  * Compatibility abstraction which allows us to generalize access to both the
@@ -44,13 +44,12 @@ public abstract class FragmentCompat<
 
   static {
     sHasSupportFragment = ReflectionUtil.tryGetClassForName(
-        "android.support.v4.app.Fragment") != null;
+        "androidx.fragment.app.Fragment") != null;
   }
 
   @Nullable
   public static FragmentCompat getFrameworkInstance() {
-    if (sFrameworkInstance == null &&
-        Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
+    if (sFrameworkInstance == null) {
       sFrameworkInstance = new FragmentCompatFramework();
     }
     return sFrameworkInstance;
