@@ -23,9 +23,6 @@ import com.facebook.stetho.inspector.elements.StyleAccumulator;
 import com.facebook.stetho.inspector.elements.StyleRuleNameAccumulator;
 import com.facebook.stetho.inspector.helper.IntegerFormatter;
 
-import javax.annotation.Nullable;
-import javax.annotation.concurrent.GuardedBy;
-
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -36,6 +33,9 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.regex.Pattern;
+
+import javax.annotation.Nullable;
+import javax.annotation.concurrent.GuardedBy;
 
 final class ViewDescriptor extends AbstractChainedDescriptor<View>
     implements HighlightableDescriptor<View> {
@@ -51,7 +51,7 @@ final class ViewDescriptor extends AbstractChainedDescriptor<View>
 
   static {
     sHasSupportNodeInfo = ReflectionUtil.tryGetClassForName(
-        "android.support.v4.view.accessibility.AccessibilityNodeInfoCompat") != null;
+        "androidx.core.view.accessibility.AccessibilityNodeInfoCompat") != null;
   }
 
   /**
