@@ -7,9 +7,6 @@
 
 package com.facebook.stetho.common.android;
 
-import android.support.annotation.Nullable;
-import android.support.v4.view.ViewCompat;
-import android.support.v4.view.accessibility.AccessibilityNodeInfoCompat;
 import android.text.TextUtils;
 import android.view.View;
 import android.view.ViewGroup;
@@ -20,6 +17,10 @@ import android.widget.ScrollView;
 import android.widget.Spinner;
 
 import java.util.List;
+
+import androidx.annotation.Nullable;
+import androidx.core.view.ViewCompat;
+import androidx.core.view.accessibility.AccessibilityNodeInfoCompat;
 
 /**
  * This class provides utility methods for determining certain accessibility properties of
