@@ -9,13 +9,14 @@ package com.facebook.stetho.common.android;
 
 import android.app.Dialog;
 import android.content.res.Resources;
-import android.support.v4.app.DialogFragment;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentManager;
 import android.view.View;
 
 import javax.annotation.Nullable;
+
+import androidx.fragment.app.DialogFragment;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentManager;
 
 final class FragmentCompatSupportLib
     extends FragmentCompat<Fragment, DialogFragment, FragmentManager, FragmentActivity> {
