@@ -8,24 +8,24 @@
 package com.facebook.stetho.dumpapp.plugins;
 
 import android.annotation.SuppressLint;
-import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.os.Build;
 import android.text.TextUtils;
+
 import com.facebook.stetho.dumpapp.DumpUsageException;
 import com.facebook.stetho.dumpapp.DumperContext;
 import com.facebook.stetho.dumpapp.DumperPlugin;
 import com.facebook.stetho.inspector.domstorage.SharedPreferencesHelper;
 
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import java.io.File;
 import java.io.PrintStream;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
 public class SharedPreferencesDumperPlugin implements DumperPlugin {
 
@@ -118,7 +118,6 @@ public class SharedPreferencesDumperPlugin implements DumperPlugin {
     return nextArg(iter, "Expected <value>");
   }
 
-  @TargetApi(Build.VERSION_CODES.HONEYCOMB)
   private static void putStringSet(
       SharedPreferences.Editor editor,
       String key,
