@@ -7,7 +7,6 @@
 
 package com.facebook.stetho.inspector.elements.android.window;
 
-import android.support.annotation.NonNull;
 import android.view.View;
 
 import java.lang.reflect.Field;
@@ -16,6 +15,8 @@ import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+
+import androidx.annotation.NonNull;
 
 class WindowRootViewCompactV18Impl extends WindowRootViewCompat {
 
