@@ -9,8 +9,9 @@ package com.facebook.stetho.inspector.helper;
 
 import android.annotation.TargetApi;
 import android.os.Build;
-import android.support.annotation.Nullable;
 import android.view.ViewDebug;
+
+import androidx.annotation.Nullable;
 
 public class IntegerFormatter {
   private static IntegerFormatter cachedFormatter;
